requirements.txt
psycopg2
requests
python-dotenv
pre-commit
python-dateutil
pytest-cov
